+++ src/OCV__LoadLibrary.java
@@ -25,7 +25,7 @@ import org.opencv.core.Point;
 import org.opencv.videoio.VideoCapture;
 
 public class OCV__LoadLibrary implements ExtendedPlugInFilter {
-    public static final String VERSION = "0.9.50.0";
+    public static final String VERSION = "0.9.51.0";
     public static final String URL_HELP = "https://github.com/WAKU-TAKE-A/IJToolsUsingOpenCV";
 
     private static boolean disposed = true;